feat(bao-io-dispatcher): add Bao Hypervisor I/O Dispatcher
The Bao I/O Dispatcher system is a kernel module that can be
loaded into the Linux kernel of the backend VMs.
It is responsible for establishing the connection between the
Remote I/O system (implemented in the Bao Hypervisor) and the
Frontend Device Model (implemented in userspace), offering a
unified API to support various VirtIO backends.

Signed-off-by: joaopeixoto13 <joaopeixotooficial@gmail.com>